Rewrite Toast
