+++ packages/bento-design-system/src/Toast/ToastContainer.tsx
@@ -1,0 +1,11 @@
+import { Box } from "..";
+import { Toast, ToastProps } from "./Toast";
+import { container } from "./ToastContainer.css";
+
+export function ToastContainer(props: ToastProps) {
+  return (
+    <Box className={container}>
+      <Toast {...props} />
+    </Box>
+  );
+}
