--- packages/bento-design-system/src/Toast/createToastInternal.tsx
@@ -1,73 +1,0 @@
-import { FunctionComponent } from "react";
-import {
-  Bleed,
-  Box,
-  Column,
-  Columns,
-  Body,
-  ButtonProps,
-  IconButtonProps,
-  ToastProps,
-  useDefaultMessages,
-} from "..";
-import { toastRecipe } from "./Toast.css";
-import { ToastConfig } from "./Config";
-
-/**
- * This component is not meant to be used directly: you should use the `showToast` function
- * provided by `useToast` instead.
- */
-export function createToastInternal(
-  config: { [P in keyof ToastConfig]-?: ToastConfig[P] },
-  {
-    Button,
-    IconButton,
-  }: {
-    Button: FunctionComponent<ButtonProps>;
-    IconButton: FunctionComponent<IconButtonProps>;
-  }
-) {
-  return function Toast({ kind, message, action, onDismiss }: ToastProps) {
-    const { defaultMessages } = useDefaultMessages();
-    return (
-      <Box
-        as="aside"
-        aria-live="polite"
-        className={toastRecipe({ kind, hasOutline: config.outline, elevation: config.elevation })}
-        borderRadius={config.radius}
-        paddingY={config.paddingY}
-        paddingX={config.paddingX}
-      >
-        <Columns space={config.internalSpacing} alignY="center">
-          <Body size={config.messageSize} color={kind === "secondary" ? "default" : kind}>
-            {message}
-          </Body>
-          {action && (
-            <Column width="content">
-              <Bleed spaceY={config.smallButtonPaddingY}>
-                <Button
-                  size={config.buttonSize}
-                  kind={config.buttonKind}
-                  hierarchy="secondary"
-                  {...action}
-                />
-              </Bleed>
-            </Column>
-          )}
-          {onDismiss && (
-            <Column width="content">
-              <IconButton
-                label={defaultMessages.Banner.dismissButtonLabel}
-                onPress={onDismiss}
-                size={config.closeIconSize}
-                kind="transparent"
-                hierarchy="secondary"
-                icon={config.closeIcon}
-              />
-            </Column>
-          )}
-        </Columns>
-      </Box>
-    );
-  };
-}
