+++ packages/bento-design-system/src/index.ts
@@ -54,7 +54,8 @@ export * from "./Stepper/Stepper";
 export * from "./Switch/Switch";
 export * from "./Table/Table";
 export * from "./Tabs/Tabs";
-export * from "./Toast/createToast";
+export * from "./Toast/Toast";
+export * from "./Toast/ToastProvider";
 export * from "./Toast/useToast";
 export * from "./Tooltip/Tooltip";
 export * from "./Typography/Body/Body";
